Clean up Avatar implementation
