+++ packages/palette/src/elements/Avatar/Avatar.ios.tsx
@@ -1,0 +1,18 @@
+import React, { FunctionComponent } from "react"
+import { styled } from "../../platform/primitives"
+import { AvatarProps, BaseAvatar, sizeValue } from "./Avatar.shared"
+
+export const Avatar: FunctionComponent<AvatarProps> = ({ ...props }) => {
+  return (
+    <BaseAvatar renderAvatar={() => <AvatarImage {...props} />} {...props} />
+  )
+}
+
+/**
+ * A circular avatar image component.
+ */
+const AvatarImage = styled.Image<AvatarProps>`
+  width: ${props => sizeValue(props).diameter};
+  height: ${props => sizeValue(props).diameter};
+  border-radius: ${props => sizeValue(props).diameter};
+`
